+++ back/src/controllers/auth/register.ts
@@ -1,12 +1,20 @@
 import { Middleware, Context } from 'koa'
 import { prisma } from '../../prisma/client'
-import { NotFoundError } from '../../helpers/errors'
+import { NotFoundError, ValidationError } from '../../helpers/errors'
 import { UserRegister } from '../../schemas/users/userRegisterSchema'
 import { processMedia } from '../../helpers/processMedia'
+import { handleSSO } from '../../helpers/handleSso'
 
 export const registerController: Middleware = async (ctx: Context) => {
-  const { dni, password, name, email, specialization }: UserRegister =
-    ctx.request.body
+  const {
+    dni,
+    password,
+    name,
+    email,
+    confirmPassword,
+    specialization,
+    itineraryId,
+  }: UserRegister = ctx.request.body
 
   const media = ctx.file
 
@@ -18,23 +26,25 @@ export const registerController: Middleware = async (ctx: Context) => {
     throw new NotFoundError('Category not found')
   }
 
+  const { status, data } = await handleSSO('register', {
+    dni,
+    password,
+    confirmPassword,
+    email,
+    itineraryId,
+  })
+  if (status !== 200) {
+    // eslint-disable-next-line @typescript-eslint/no-throw-literal
+    throw new ValidationError(data.message)
+  }
+
   const user = await prisma.user.create({
     data: {
-      dni: dni.toUpperCase(),
-      password,
+      id: data.id,
       name,
-      email,
       specializationId: existingCategory.id,
     },
   })
-
-  if (!user || user.dni !== dni.toUpperCase()) {
-    ctx.status = 500
-    ctx.body = {
-      error: 'Database error',
-    }
-    return
-  }
 
   if (media) {
     const { mediaId } = await processMedia(media, user.id)