+++ sso/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.4.0] - 2023-12-14
+
+### Changed
+
+- Register, Validate, Login endpoint returns user id
+- InvalidParamError to catch foreign_key error
+- EmailError to EmailDniError
+
+### Added
+
+- DuplicateError to catch unique_key error in errorMiddleware
+
+### Testing
+
+- Added tests for GET itinerary.
+- Updated test for POST register.
+
+### Documentation
+
+- Updated Login and Validate endpoints
+- Added duplicateError response and register response for POST register user.
+  
 ## [1.3.0] - 2023-11-23
 
 ### Added